Criado camada de serviços e registro de componente nessa classe usando @Service. Método findAll e findById

+++ src/main/java/com/blinkspace/springjpa/repositories/UserRepository.java
@@ -3,5 +3,6 @@ package com.blinkspace.springjpa.repositories;
 import com.blinkspace.springjpa.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+//não precisa colocar @Repository nesse caso, pq já esta herdadndo da interface do JpaRepository, que já tem a anotação
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
